Replace ingredients page with products page
This consolidates product listings with ingredients information into a single page.
