+++ cypress/types/route-path.ts
@@ -1,8 +1,8 @@
 export enum RoutePath {
   About = '/about',
   Home = '/',
-  Ingredients = '/ingredients',
   NotFound = '/not-found',
+  Products = '/products',
   Purchase = '/purchase',
   Recipes = '/recipes',
 }
