+++ src/components/Products.tsx
@@ -1,21 +1,28 @@
 import React from 'react';
 import ProductListing from './ProductListing';
 import { PRODUCTS } from '../constants/product';
+import { Helmet } from 'react-helmet-async';
+import styled from 'styled-components';
+import Typography from './Typography';
 
-interface Props {
-  onClickOrderOnline: () => void;
-}
+const StyledUnorderedList = styled.ul`
+  padding: 0;
+  list-style: none;
+  margin: 0;
+`;
 
-const Products: React.FC<Props> = ({ onClickOrderOnline }) => {
+const Products: React.FC = () => {
   return (
     <>
-      {PRODUCTS.map((product) => (
-        <ProductListing
-          key={product.title}
-          product={product}
-          onClickOrderOnline={onClickOrderOnline}
-        />
-      ))}
+      <Helmet>
+        <title>Products · Jerrypop</title>
+      </Helmet>
+      <Typography type="h1">Products</Typography>
+      <StyledUnorderedList>
+        {PRODUCTS.map((product) => (
+          <ProductListing key={product.title} product={product} />
+        ))}
+      </StyledUnorderedList>
     </>
   );
 };
