--- src/constants/ingredients.ts
@@ -1,97 +1,0 @@
-import { IngredientDetails } from '../types/ingredients';
-
-export const INGREDIENTS: IngredientDetails[] = [
-  {
-    allergens: ['MILK'],
-    ingredients: [
-      'POPCORN',
-      'CANOLA OIL',
-      'BUTTERMILK POWDER',
-      'ONION POWDER',
-      'GARLIC POWDER',
-      'SALT',
-      'SUGAR',
-      'MONOSODIUM GLUTAMATE',
-      'GROUND HABANERO PEPPER',
-      'DILL',
-      'BLACK PEPPERCORN',
-      'CHIVES',
-      'PARSLEY',
-    ],
-    title: 'Habanero Ranch',
-  },
-  {
-    allergens: [],
-    ingredients: [
-      'POPCORN',
-      'CANOLA OIL',
-      'EXTRA VIRGIN OLIVE OIL',
-      'GRANULATED GARLIC',
-      'GRANULATED ONION',
-      'SALT',
-      '100% NATURAL LEMON POWDER',
-      'SUGAR',
-      'ROSEMARY',
-      'BLACK PEPPERCORN',
-    ],
-    title: 'Rosemary Lemon Pepper',
-  },
-  {
-    allergens: ['FISH', 'MILK'],
-    ingredients: [
-      'POPCORN',
-      'CANOLA OIL',
-      'NORI KOMI FURIKAKE (SESAME SEED, SUGAR, SEAWEED, SALT, MALTODEXTRIN, DISODIUM SUCCINATE, DISODIUM INOSINATE)',
-      'SALT',
-      'BONITO SOUP STOCK (SALT, MONOSODIUM GLUTAMATE, LACTOSE (MILK), SUGAR, DRIED BONITO TUNA, DISODIUM INOSINATE, YEAST EXTRACT, BONITO TUNA BROTH, DISODIUM SUCCINATE, BONITO TUNA EXTRACT, XYLOSE, MALTOSE, DEXTROSE)',
-      'MONOSODIUM GLUTAMATE',
-      'SHICHIMI TOGARASHI (CHILI PEPPER, ORANGE PEEL, BLACK SESAME SEED, SEAWEED, WHITE SESAME SEED, JAPANESE PEPPER, GINGER)',
-      'GRANULATED ONION',
-      'SUGAR',
-    ],
-    title: 'Furikake Dashi',
-  },
-  {
-    allergens: [],
-    ingredients: [
-      'POPCORN',
-      'CANOLA OIL',
-      'NUTRITIONAL YEAST (Dried Yeast, Niacin, Pyridoxine Hydrochloride (Vitamin B6), Thiamin Hydrochloride (Vitamin B1), Riboflavin (Vitamin B2), Folic Acid (Vitamin B9), Cyanocobalamin (Vitamin B12))',
-      'GRANULATED GARLIC',
-      'GRANULATED ONION',
-      'SALT',
-      'SICHUAN PEPPERCORN',
-      'MONOSODIUM GLUTAMATE',
-      'SUGAR',
-      'CRYSTALLIZED LIME (CITRIC ACID, LIME OIL, LIME JUICE, MALTODEXTRIN)',
-    ],
-    title: 'Sichuan Nooch',
-  },
-  {
-    allergens: ['PEANUTS', 'SOY'],
-    ingredients: [
-      'STRAWBERRY JAM',
-      'PEANUT BUTTER (ROASTED PEANUTS, SUGAR, HYDROGENATED VEGETABLE OIL (COTTONSEED, SOYBEAN AND RAPESEED OIL), SALT)',
-      'POPCORN',
-      'CANOLA OIL',
-      'NATURALLY REFINED ORGANIC COCONUT OIL',
-      'SUGAR',
-      'BAKING POWDER',
-      'SALT',
-    ],
-    title: 'PB&J',
-  },
-  {
-    allergens: ['TREE NUTS (HAZELNUTS)', 'MILK', 'SOY'],
-    ingredients: [
-      'HAZELNUT SPREAD WITH COCOA (SUGAR, PALM OIL, HAZELNUTS, SKIM MILK, COCOA, LECITHIN, VANILLIN)',
-      'POPCORN',
-      'CANOLA OIL',
-      'NATURALLY REFINED ORGANIC COCONUT OIL',
-      'SUGAR',
-      'BAKING POWDER',
-      'SALT',
-    ],
-    title: 'Chocolate Hazelnut Espresso',
-  },
-];
