--- src/types/ingredients.ts
@@ -1,5 +1,0 @@
-export interface IngredientDetails {
-  allergens: string[];
-  ingredients: string[];
-  title: string;
-}
